requirements.txt
numpy
pandas
scrapy
fastapi
uvicorn
Jinja2
appwrite
python-dotenv
pymongo
aiohttp
motor
httpx
numpy
sentence-transformers
scikit-learn
-e .